+++ client/src/components/AuthRoute.tsx
@@ -4,6 +4,7 @@ import State from "../interfaces/State";
 import { Redirect, Route, RouteComponentProps, useHistory, useLocation } from "react-router-dom";
 import User from "../interfaces/User";
 import Loader from "./loader";
+import CadInfo from "../interfaces/CadInfo";
 
 interface Props {
   Component: any;
@@ -12,13 +13,37 @@ interface Props {
   path: string;
   requirement?: "admin" | "leo" | "dispatch" | "tow" | "ems_fd" | "supervisor";
   user: User | null;
+  cadInfo: CadInfo | null;
 }
 
 export const adminRanks: string[] = ["owner", "admin", "moderator"];
 
-const AuthRoute: React.FC<Props> = ({ Component, loading, isAuth, path, user, requirement }) => {
+const AuthRoute: React.FC<Props> = ({
+  Component,
+  loading,
+  isAuth,
+  path,
+  user,
+  requirement,
+  cadInfo,
+}) => {
   const history = useHistory();
   const location = useLocation();
+
+  React.useEffect(() => {
+    const parsedPath = path
+      .split("")
+      .filter((v) => v !== "/")
+      .join("")
+      .toLowerCase();
+
+    if (
+      ["bleeter", "tow", "truck-logs", "courthouse", "taxi"].includes(parsedPath) &&
+      !cadInfo?.features.includes(parsedPath)
+    ) {
+      return history.push("/not-enabled");
+    }
+  }, [cadInfo?.features, history, path]);
 
   React.useEffect(() => {
     if (requirement && !loading && isAuth) {
@@ -93,6 +118,7 @@ const mapToProps = (state: State) => ({
   user: state.auth.user,
   isAuth: state.auth.isAuth,
   loading: state.auth.loading,
+  cadInfo: state.global.cadInfo,
 });
 
 export default connect(mapToProps)(AuthRoute);