+++ client/src/index.tsx
@@ -12,6 +12,7 @@ import store from "./lib/store";
 const HomePage = React.lazy(() => import("./pages/index"));
 const NotFoundPage = React.lazy(() => import("./pages/not-found"));
 const ForbiddenPage = React.lazy(() => import("./pages/forbidden"));
+const NotEnabledPage = React.lazy(() => import("./pages/not-enabled"));
 
 const BleeterPage = React.lazy(() => import("./pages/bleeter"));
 const BleetPage = React.lazy(() => import("./pages/bleeter/bleet"));
@@ -120,7 +121,8 @@ ReactDOM.render(
             <AuthRoute requirement="dispatch" path="/dispatch/map" Component={Map} />
             <AuthRoute requirement="dispatch" path="/dispatch" Component={DispatchDash} />
 
-            <AuthRoute path="/court" Component={CourthousePage} />
+            <Route path="/court" render={() => (window.location.pathname = "/courthouse")} />
+            <AuthRoute path="/courthouse" Component={CourthousePage} />
 
             <AuthRoute
               path="/company/:citizenId/:companyId/manage/:id"
@@ -243,6 +245,7 @@ ReactDOM.render(
 
             <Route path="/logout" component={LogoutPage} />
             <Route path="/forbidden" component={ForbiddenPage} />
+            <Route path="/not-enabled" component={NotEnabledPage} />
             <Route component={NotFoundPage} />
           </Switch>
         </React.Suspense>